% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-mutiples.R
\name{add_to_mrset}
\alias{add_to_mrset}
\title{Adds a value to a multiple-response set, ensuring uniqueness and order.}
\usage{
add_to_mrset(vec, value)
}
\description{
Adds a value to a multiple-response set, ensuring uniqueness and order.
}
