% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ds.R
\name{base}
\alias{base}
\title{Selects variables named PREFIX_1, PREFIX_2, ..., for one or more literal prefixes.}
\usage{
base(...)
}
\description{
Selects variables named PREFIX_1, PREFIX_2, ..., for one or more literal prefixes.
}
