% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds-vars-to-cases.R
\name{cases_index}
\alias{cases_index}
\title{Describes the index variable created by \verb{$clone_to_cases()}.}
\usage{
cases_index(name, label = NULL, val_labels = NULL)
}
\description{
Describes the index variable created by \verb{$clone_to_cases()}.
}
