% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp2.R
\name{DS}
\alias{DS}
\title{DS Class}
\description{
The `DS` class manages both raw data and its associated metadata, facilitating streamlined data manipulation.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{A tibble containing the dataset, with each column representing a different variable (e.g., age, gender, survey responses).}

\item{\code{var_labels}}{A named list mapping each variable's identifier to a descriptive label, enhancing readability in outputs and reports.}

\item{\code{val_labels}}{A named list for categorical variables, where each entry maps numeric codes to meaningful category labels (e.g., 1 = "Agree", 2 = "Disagree").}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DS-export_spss}{\code{DS$export_spss()}}
\item \href{#method-DS-calc_funnel}{\code{DS$calc_funnel()}}
\item \href{#method-DS-funnel_tables}{\code{DS$funnel_tables()}}
\item \href{#method-DS-merge_data}{\code{DS$merge_data()}}
\item \href{#method-DS-open_export}{\code{DS$open_export()}}
\item \href{#method-DS-open_import}{\code{DS$open_import()}}
\item \href{#method-DS-qtable}{\code{DS$qtable()}}
\item \href{#method-DS-flip_scale}{\code{DS$flip_scale()}}
\item \href{#method-DS-scale_flip}{\code{DS$scale_flip()}}
\item \href{#method-DS-scale_dense}{\code{DS$scale_dense()}}
\item \href{#method-DS-scale_shift}{\code{DS$scale_shift()}}
\item \href{#method-DS-scale_move}{\code{DS$scale_move()}}
\item \href{#method-DS-recode_from_to}{\code{DS$recode_from_to()}}
\item \href{#method-DS-make_tb}{\code{DS$make_tb()}}
\item \href{#method-DS-make_t2b}{\code{DS$make_t2b()}}
\item \href{#method-DS-make_t3b}{\code{DS$make_t3b()}}
\item \href{#method-DS-prepare_val_labels}{\code{DS$prepare_val_labels()}}
\item \href{#method-DS-calc_table}{\code{DS$calc_table()}}
\item \href{#method-DS-vars_transpose}{\code{DS$vars_transpose()}}
\item \href{#method-DS-autocode_single}{\code{DS$autocode_single()}}
\item \href{#method-DS-merge_vars}{\code{DS$merge_vars()}}
\item \href{#method-DS-to_multiple}{\code{DS$to_multiple()}}
\item \href{#method-DS-to_single}{\code{DS$to_single()}}
\item \href{#method-DS-recode}{\code{DS$recode()}}
\item \href{#method-DS-transfer}{\code{DS$transfer()}}
\item \href{#method-DS-recode_empty}{\code{DS$recode_empty()}}
\item \href{#method-DS-recalc_empty}{\code{DS$recalc_empty()}}
\item \href{#method-DS-compute}{\code{DS$compute()}}
\item \href{#method-DS-mutate}{\code{DS$mutate()}}
\item \href{#method-DS-nvn}{\code{DS$nvn()}}
\item \href{#method-DS-nvs}{\code{DS$nvs()}}
\item \href{#method-DS-nvm}{\code{DS$nvm()}}
\item \href{#method-DS-add_total}{\code{DS$add_total()}}
\item \href{#method-DS-add_rid}{\code{DS$add_rid()}}
\item \href{#method-DS-nvn_src}{\code{DS$nvn_src()}}
\item \href{#method-DS-nvs_src}{\code{DS$nvs_src()}}
\item \href{#method-DS-transfer_to}{\code{DS$transfer_to()}}
\item \href{#method-DS-nvclone_to}{\code{DS$nvclone_to()}}
\item \href{#method-DS-set_if}{\code{DS$set_if()}}
\item \href{#method-DS-set_na_if}{\code{DS$set_na_if()}}
\item \href{#method-DS-add_if}{\code{DS$add_if()}}
\item \href{#method-DS-add_net}{\code{DS$add_net()}}
\item \href{#method-DS-vdiscard}{\code{DS$vdiscard()}}
\item \href{#method-DS-vstrip}{\code{DS$vstrip()}}
\item \href{#method-DS-nvclone}{\code{DS$nvclone()}}
\item \href{#method-DS-make_means}{\code{DS$make_means()}}
\item \href{#method-DS-make_nps0_groups}{\code{DS$make_nps0_groups()}}
\item \href{#method-DS-make_nps1_groups}{\code{DS$make_nps1_groups()}}
\item \href{#method-DS-make_nps0_scores}{\code{DS$make_nps0_scores()}}
\item \href{#method-DS-make_nps1_scores}{\code{DS$make_nps1_scores()}}
\item \href{#method-DS-get_spss}{\code{DS$get_spss()}}
\item \href{#method-DS-get_rds}{\code{DS$get_rds()}}
\item \href{#method-DS-new}{\code{DS$new()}}
\item \href{#method-DS-save}{\code{DS$save()}}
\item \href{#method-DS-get_col_names}{\code{DS$get_col_names()}}
\item \href{#method-DS-names}{\code{DS$names()}}
\item \href{#method-DS-base_name}{\code{DS$base_name()}}
\item \href{#method-DS-var_type}{\code{DS$var_type()}}
\item \href{#method-DS-is_nominal}{\code{DS$is_nominal()}}
\item \href{#method-DS-filter}{\code{DS$filter()}}
\item \href{#method-DS-keep}{\code{DS$keep()}}
\item \href{#method-DS-remove}{\code{DS$remove()}}
\item \href{#method-DS-rename}{\code{DS$rename()}}
\item \href{#method-DS-move}{\code{DS$move()}}
\item \href{#method-DS-clone_if}{\code{DS$clone_if()}}
\item \href{#method-DS-vacuum}{\code{DS$vacuum()}}
\item \href{#method-DS-get_var_label}{\code{DS$get_var_label()}}
\item \href{#method-DS-get_var_labels}{\code{DS$get_var_labels()}}
\item \href{#method-DS-add_label_suffix}{\code{DS$add_label_suffix()}}
\item \href{#method-DS-set_var_label}{\code{DS$set_var_label()}}
\item \href{#method-DS-set_val_labels}{\code{DS$set_val_labels()}}
\item \href{#method-DS-set_labels}{\code{DS$set_labels()}}
\item \href{#method-DS-add_labels}{\code{DS$add_labels()}}
\item \href{#method-DS-remove_labels}{\code{DS$remove_labels()}}
\item \href{#method-DS-vars_to_cases}{\code{DS$vars_to_cases()}}
\item \href{#method-DS-set_multiples}{\code{DS$set_multiples()}}
\item \href{#method-DS-conv_multiples}{\code{DS$conv_multiples()}}
\item \href{#method-DS-var_view}{\code{DS$var_view()}}
\item \href{#method-DS-var_check}{\code{DS$var_check()}}
\item \href{#method-DS-export_copy}{\code{DS$export_copy()}}
\item \href{#method-DS-clone}{\code{DS$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-export_spss"></a>}}
\if{latex}{\out{\hypertarget{method-DS-export_spss}{}}}
\subsection{Method \code{export_spss()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$export_spss(sav_name, sps_name, CP1251 = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-calc_funnel"></a>}}
\if{latex}{\out{\hypertarget{method-DS-calc_funnel}{}}}
\subsection{Method \code{calc_funnel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$calc_funnel(
  ...,
  weight = NULL,
  exclude_codes = NULL,
  sheet = "",
  filename = NULL,
  caption = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-funnel_tables"></a>}}
\if{latex}{\out{\hypertarget{method-DS-funnel_tables}{}}}
\subsection{Method \code{funnel_tables()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$funnel_tables(..., filename = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-merge_data"></a>}}
\if{latex}{\out{\hypertarget{method-DS-merge_data}{}}}
\subsection{Method \code{merge_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$merge_data(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-open_export"></a>}}
\if{latex}{\out{\hypertarget{method-DS-open_export}{}}}
\subsection{Method \code{open_export()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$open_export(
  key,
  ...,
  add = NULL,
  type = "multiple",
  filename = "open temp",
  sheet = NULL,
  alias = NULL,
  labels = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-open_import"></a>}}
\if{latex}{\out{\hypertarget{method-DS-open_import}{}}}
\subsection{Method \code{open_import()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$open_import(path, sheets = NULL, after = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-qtable"></a>}}
\if{latex}{\out{\hypertarget{method-DS-qtable}{}}}
\subsection{Method \code{qtable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$qtable(var, weight = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-flip_scale"></a>}}
\if{latex}{\out{\hypertarget{method-DS-flip_scale}{}}}
\subsection{Method \code{flip_scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$flip_scale(vars, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-scale_flip"></a>}}
\if{latex}{\out{\hypertarget{method-DS-scale_flip}{}}}
\subsection{Method \code{scale_flip()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$scale_flip(vars, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-scale_dense"></a>}}
\if{latex}{\out{\hypertarget{method-DS-scale_dense}{}}}
\subsection{Method \code{scale_dense()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$scale_dense(vars)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-scale_shift"></a>}}
\if{latex}{\out{\hypertarget{method-DS-scale_shift}{}}}
\subsection{Method \code{scale_shift()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$scale_shift(vars, from, amount = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-scale_move"></a>}}
\if{latex}{\out{\hypertarget{method-DS-scale_move}{}}}
\subsection{Method \code{scale_move()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$scale_move(vars, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-recode_from_to"></a>}}
\if{latex}{\out{\hypertarget{method-DS-recode_from_to}{}}}
\subsection{Method \code{recode_from_to()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$recode_from_to(from_vars, to_vars, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-make_tb"></a>}}
\if{latex}{\out{\hypertarget{method-DS-make_tb}{}}}
\subsection{Method \code{make_tb()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$make_tb(
  vars,
  box_size = 2,
  ...,
  suffix = NULL,
  label_suffix = NULL,
  labels = NULL,
  is_reverse = F,
  range = NULL,
  move = T,
  suffix_position = "auto"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-make_t2b"></a>}}
\if{latex}{\out{\hypertarget{method-DS-make_t2b}{}}}
\subsection{Method \code{make_t2b()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$make_t2b(
  vars,
  ...,
  suffix = "T2B",
  label_suffix = "(T2B)",
  labels = NULL,
  is_reverse = F,
  range = NULL,
  move = T,
  suffix_position = "auto"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-make_t3b"></a>}}
\if{latex}{\out{\hypertarget{method-DS-make_t3b}{}}}
\subsection{Method \code{make_t3b()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$make_t3b(
  vars,
  ...,
  suffix = "T3B",
  label_suffix = "(T3B)",
  labels = NULL,
  is_reverse = F,
  range = NULL,
  move = T,
  suffix_position = "auto"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-prepare_val_labels"></a>}}
\if{latex}{\out{\hypertarget{method-DS-prepare_val_labels}{}}}
\subsection{Method \code{prepare_val_labels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$prepare_val_labels(var, warning = F)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-calc_table"></a>}}
\if{latex}{\out{\hypertarget{method-DS-calc_table}{}}}
\subsection{Method \code{calc_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$calc_table(
  row_vars,
  col_vars = NULL,
  weight = NULL,
  waves = NULL,
  sig = "first",
  sheet = "",
  filename = NULL,
  open = F
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-vars_transpose"></a>}}
\if{latex}{\out{\hypertarget{method-DS-vars_transpose}{}}}
\subsection{Method \code{vars_transpose()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$vars_transpose(vars, new_name, label_prefix, na_code, remove_code)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-autocode_single"></a>}}
\if{latex}{\out{\hypertarget{method-DS-autocode_single}{}}}
\subsection{Method \code{autocode_single()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$autocode_single(..., labels = NULL, nomatch = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-merge_vars"></a>}}
\if{latex}{\out{\hypertarget{method-DS-merge_vars}{}}}
\subsection{Method \code{merge_vars()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$merge_vars(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-to_multiple"></a>}}
\if{latex}{\out{\hypertarget{method-DS-to_multiple}{}}}
\subsection{Method \code{to_multiple()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$to_multiple(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-to_single"></a>}}
\if{latex}{\out{\hypertarget{method-DS-to_single}{}}}
\subsection{Method \code{to_single()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$to_single(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-recode"></a>}}
\if{latex}{\out{\hypertarget{method-DS-recode}{}}}
\subsection{Method \code{recode()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$recode(vars, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-transfer"></a>}}
\if{latex}{\out{\hypertarget{method-DS-transfer}{}}}
\subsection{Method \code{transfer()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$transfer(vars, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-recode_empty"></a>}}
\if{latex}{\out{\hypertarget{method-DS-recode_empty}{}}}
\subsection{Method \code{recode_empty()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$recode_empty(cols, value, label = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-recalc_empty"></a>}}
\if{latex}{\out{\hypertarget{method-DS-recalc_empty}{}}}
\subsection{Method \code{recalc_empty()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$recalc_empty(cols, value, label = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-compute"></a>}}
\if{latex}{\out{\hypertarget{method-DS-compute}{}}}
\subsection{Method \code{compute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$compute(var_name, expr)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-mutate"></a>}}
\if{latex}{\out{\hypertarget{method-DS-mutate}{}}}
\subsection{Method \code{mutate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$mutate(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-nvn"></a>}}
\if{latex}{\out{\hypertarget{method-DS-nvn}{}}}
\subsection{Method \code{nvn()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$nvn(
  name,
  label = NULL,
  labels = NULL,
  fill = NA,
  after = NULL,
  before = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-nvs"></a>}}
\if{latex}{\out{\hypertarget{method-DS-nvs}{}}}
\subsection{Method \code{nvs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$nvs(name, label = NULL, fill = NA, after = NULL, before = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-nvm"></a>}}
\if{latex}{\out{\hypertarget{method-DS-nvm}{}}}
\subsection{Method \code{nvm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$nvm(name, label = NULL, labels = NULL, after = NULL, before = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-add_total"></a>}}
\if{latex}{\out{\hypertarget{method-DS-add_total}{}}}
\subsection{Method \code{add_total()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$add_total()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-add_rid"></a>}}
\if{latex}{\out{\hypertarget{method-DS-add_rid}{}}}
\subsection{Method \code{add_rid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$add_rid(
  name = "RID",
  label = NULL,
  fill = row_number(),
  after = NULL,
  before = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-nvn_src"></a>}}
\if{latex}{\out{\hypertarget{method-DS-nvn_src}{}}}
\subsection{Method \code{nvn_src()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$nvn_src(
  vars,
  suffix = NULL,
  label_suffix = NULL,
  labels = NULL,
  move = T,
  presuffix = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-nvs_src"></a>}}
\if{latex}{\out{\hypertarget{method-DS-nvs_src}{}}}
\subsection{Method \code{nvs_src()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$nvs_src(
  vars,
  suffix = NULL,
  label_suffix = NULL,
  move = T,
  presuffix = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-transfer_to"></a>}}
\if{latex}{\out{\hypertarget{method-DS-transfer_to}{}}}
\subsection{Method \code{transfer_to()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$transfer_to(new_vars, from_vars, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-nvclone_to"></a>}}
\if{latex}{\out{\hypertarget{method-DS-nvclone_to}{}}}
\subsection{Method \code{nvclone_to()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$nvclone_to(new_var, from_var, label = NULL, after = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-set_if"></a>}}
\if{latex}{\out{\hypertarget{method-DS-set_if}{}}}
\subsection{Method \code{set_if()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$set_if(var, value, ..., label = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-set_na_if"></a>}}
\if{latex}{\out{\hypertarget{method-DS-set_na_if}{}}}
\subsection{Method \code{set_na_if()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$set_na_if(var, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-add_if"></a>}}
\if{latex}{\out{\hypertarget{method-DS-add_if}{}}}
\subsection{Method \code{add_if()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$add_if(var, value, ..., label = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-add_net"></a>}}
\if{latex}{\out{\hypertarget{method-DS-add_net}{}}}
\subsection{Method \code{add_net()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$add_net(vars, value, ..., label = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-vdiscard"></a>}}
\if{latex}{\out{\hypertarget{method-DS-vdiscard}{}}}
\subsection{Method \code{vdiscard()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$vdiscard(vars, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-vstrip"></a>}}
\if{latex}{\out{\hypertarget{method-DS-vstrip}{}}}
\subsection{Method \code{vstrip()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$vstrip(vars, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-nvclone"></a>}}
\if{latex}{\out{\hypertarget{method-DS-nvclone}{}}}
\subsection{Method \code{nvclone()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$nvclone(
  vars,
  ...,
  suffix = NULL,
  label_suffix = NULL,
  labels = NULL,
  move = T,
  suffix_position = "auto",
  else_copy = F
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-make_means"></a>}}
\if{latex}{\out{\hypertarget{method-DS-make_means}{}}}
\subsection{Method \code{make_means()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$make_means(
  vars,
  ...,
  suffix = "MEAN",
  label_suffix = "(MEAN)",
  move = T,
  suffix_position = "auto",
  vdiscard = NULL,
  else_copy = F
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-make_nps0_groups"></a>}}
\if{latex}{\out{\hypertarget{method-DS-make_nps0_groups}{}}}
\subsection{Method \code{make_nps0_groups()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$make_nps0_groups(
  vars,
  suffix = "GROUP",
  label_suffix = "(GROUP)",
  move = T,
  labels = c("Detractors", "Neutrals", "Promoters"),
  suffix_position = "auto"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-make_nps1_groups"></a>}}
\if{latex}{\out{\hypertarget{method-DS-make_nps1_groups}{}}}
\subsection{Method \code{make_nps1_groups()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$make_nps1_groups(
  vars,
  suffix = "GROUP",
  label_suffix = "(GROUP)",
  move = T,
  labels = c("Detractors", "Neutrals", "Promoters"),
  suffix_position = "auto"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-make_nps0_scores"></a>}}
\if{latex}{\out{\hypertarget{method-DS-make_nps0_scores}{}}}
\subsection{Method \code{make_nps0_scores()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$make_nps0_scores(
  vars,
  suffix = "SCORE",
  label_suffix = "(SCORE)",
  move = T,
  suffix_position = "auto"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-make_nps1_scores"></a>}}
\if{latex}{\out{\hypertarget{method-DS-make_nps1_scores}{}}}
\subsection{Method \code{make_nps1_scores()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$make_nps1_scores(
  vars,
  suffix = "SCORE",
  label_suffix = "(SCORE)",
  move = T,
  suffix_position = "auto"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-get_spss"></a>}}
\if{latex}{\out{\hypertarget{method-DS-get_spss}{}}}
\subsection{Method \code{get_spss()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$get_spss(filename)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-get_rds"></a>}}
\if{latex}{\out{\hypertarget{method-DS-get_rds}{}}}
\subsection{Method \code{get_rds()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$get_rds(filename)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-new"></a>}}
\if{latex}{\out{\hypertarget{method-DS-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$new(filename = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-save"></a>}}
\if{latex}{\out{\hypertarget{method-DS-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$save(filename)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-get_col_names"></a>}}
\if{latex}{\out{\hypertarget{method-DS-get_col_names}{}}}
\subsection{Method \code{get_col_names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$get_col_names(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-names"></a>}}
\if{latex}{\out{\hypertarget{method-DS-names}{}}}
\subsection{Method \code{names()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$names(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-base_name"></a>}}
\if{latex}{\out{\hypertarget{method-DS-base_name}{}}}
\subsection{Method \code{base_name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$base_name(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-var_type"></a>}}
\if{latex}{\out{\hypertarget{method-DS-var_type}{}}}
\subsection{Method \code{var_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$var_type(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-is_nominal"></a>}}
\if{latex}{\out{\hypertarget{method-DS-is_nominal}{}}}
\subsection{Method \code{is_nominal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$is_nominal(vars)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-filter"></a>}}
\if{latex}{\out{\hypertarget{method-DS-filter}{}}}
\subsection{Method \code{filter()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$filter(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-keep"></a>}}
\if{latex}{\out{\hypertarget{method-DS-keep}{}}}
\subsection{Method \code{keep()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$keep(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-remove"></a>}}
\if{latex}{\out{\hypertarget{method-DS-remove}{}}}
\subsection{Method \code{remove()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$remove(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-rename"></a>}}
\if{latex}{\out{\hypertarget{method-DS-rename}{}}}
\subsection{Method \code{rename()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$rename(names_from, names_to)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-move"></a>}}
\if{latex}{\out{\hypertarget{method-DS-move}{}}}
\subsection{Method \code{move()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$move(..., after = NULL, before = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-clone_if"></a>}}
\if{latex}{\out{\hypertarget{method-DS-clone_if}{}}}
\subsection{Method \code{clone_if()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$clone_if(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-vacuum"></a>}}
\if{latex}{\out{\hypertarget{method-DS-vacuum}{}}}
\subsection{Method \code{vacuum()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$vacuum()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-get_var_label"></a>}}
\if{latex}{\out{\hypertarget{method-DS-get_var_label}{}}}
\subsection{Method \code{get_var_label()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$get_var_label(var)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-get_var_labels"></a>}}
\if{latex}{\out{\hypertarget{method-DS-get_var_labels}{}}}
\subsection{Method \code{get_var_labels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$get_var_labels(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-add_label_suffix"></a>}}
\if{latex}{\out{\hypertarget{method-DS-add_label_suffix}{}}}
\subsection{Method \code{add_label_suffix()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$add_label_suffix(vars, suffix, sep = " ")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-set_var_label"></a>}}
\if{latex}{\out{\hypertarget{method-DS-set_var_label}{}}}
\subsection{Method \code{set_var_label()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$set_var_label(var, label)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-set_val_labels"></a>}}
\if{latex}{\out{\hypertarget{method-DS-set_val_labels}{}}}
\subsection{Method \code{set_val_labels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$set_val_labels(vars, labels)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-set_labels"></a>}}
\if{latex}{\out{\hypertarget{method-DS-set_labels}{}}}
\subsection{Method \code{set_labels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$set_labels(var, label, labels)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-add_labels"></a>}}
\if{latex}{\out{\hypertarget{method-DS-add_labels}{}}}
\subsection{Method \code{add_labels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$add_labels(vars, new_labels)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-remove_labels"></a>}}
\if{latex}{\out{\hypertarget{method-DS-remove_labels}{}}}
\subsection{Method \code{remove_labels()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$remove_labels(vars, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-vars_to_cases"></a>}}
\if{latex}{\out{\hypertarget{method-DS-vars_to_cases}{}}}
\subsection{Method \code{vars_to_cases()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$vars_to_cases(
  index,
  ...,
  index_label = NULL,
  index_values = NULL,
  index_labels = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-set_multiples"></a>}}
\if{latex}{\out{\hypertarget{method-DS-set_multiples}{}}}
\subsection{Method \code{set_multiples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$set_multiples()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-conv_multiples"></a>}}
\if{latex}{\out{\hypertarget{method-DS-conv_multiples}{}}}
\subsection{Method \code{conv_multiples()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$conv_multiples()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-var_view"></a>}}
\if{latex}{\out{\hypertarget{method-DS-var_view}{}}}
\subsection{Method \code{var_view()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$var_view(name = NULL, label = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-var_check"></a>}}
\if{latex}{\out{\hypertarget{method-DS-var_check}{}}}
\subsection{Method \code{var_check()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$var_check(name = NULL, label = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-export_copy"></a>}}
\if{latex}{\out{\hypertarget{method-DS-export_copy}{}}}
\subsection{Method \code{export_copy()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$export_copy(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DS-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DS-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
