% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{eval_row_mask}
\alias{eval_row_mask}
\title{#' Checks if elements are valid based on their type.
#' @export
is_valid = function(xs) .is_valid(xs)}
\usage{
eval_row_mask(condition, data)
}
\description{
#' Checks if elements are valid based on their type.
#' @export
is_valid = function(xs) .is_valid(xs)
}
