% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp2 has.R
\name{has}
\alias{has}
\title{Check if Values are Present in a Variable}
\usage{
has(var, ...)
}
\arguments{
\item{var}{A variable to check, which can be a list or a vector.}

\item{...}{One or more values to check for within `var`.}
}
\value{
A logical vector indicating if the specified values are found within the elements of `var`. If a single value is provided, it checks for the presence of that value. If multiple values are provided, it checks for the presence of any of those values.
}
\description{
The `has` function checks if specific values are present within the elements of a given variable. The input variable can be a list or a vector.
}
\examples{
# Check for a single value in a list
has(list(c(1, 2, 3), c(4, 5, 6)), 2)

# Check for multiple values in a list
has(list(c(1, 2, 3), c(4, 5, 6)), 2, 5)

# Check for a single value in a vector
has(c(1, 2, 3, 4, 5), 3)

# Check for multiple values in a vector
has(c(1, 2, 3, 4, 5), 3:4, 6)

}
