% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-ds.R
\name{is_multiple}
\alias{is_multiple}
\title{Checks whether a vector uses the rdp2 multiple-response representation.}
\usage{
is_multiple(x)
}
\description{
Checks whether a vector uses the rdp2 multiple-response representation.
}
