% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_present}
\alias{is_present}
\title{Check whether variable elements contain a response}
\usage{
is_present(x)
}
\arguments{
\item{x}{A variable vector or multiple-response list column.}
}
\value{
A logical vector indicating which elements contain a response.
}
\description{
This is the inverse of \code{is_empty()}.
}
