% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp2.R
\name{base}
\alias{base}
\title{Returns a tidyselect matches() expression that selects variables with the given prefixes followed by _ and digits.}
\usage{
base(...)
}
\description{
Returns a tidyselect matches() expression that selects variables with the given prefixes followed by _ and digits.
}
