% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{base}
\alias{base}
\title{Creates a regex pattern to match variables with specified prefixes followed by an underscore and digits.}
\usage{
base(...)
}
\description{
Creates a regex pattern to match variables with specified prefixes followed by an underscore and digits.
}
