% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdp2-package.R
\docType{package}
\name{rdp2-package}
\alias{rdp2}
\alias{rdp2-package}
\title{rdp2: R Data Processing}
\description{
A library for common data processing tasks and table outputs. Recoding variables, creating excel tebles,, multiple response questions support, working with weighted data support.
}
\author{
\strong{Maintainer}: Kristensen Kor \email{kristensen.kor@gmail.com}

}
\keyword{internal}
